feat: Add custom hook for managing filters and add responsive ui
